Rename drugwars package to agentics and update tests
