+++ setup.py
@@ -4,14 +4,14 @@ import sys
 from setuptools import setup, find_packages
 
 setup(
-    name='drugwars',
+    name='agentics',
     version='1.2.1-post3',
     author='Max Bridgland',
     author_email='mabridgland@protonmail.com',
     description='The 80s DOS game re-written in Python',
     long_description=open('./README.md', 'r').read(),
     long_description_content_type='text/markdown',
-    url='https://github.com/M4cs/Drugwars',
+    url='https://github.com/Rolphs/TheStockWolf',
     packages=find_packages(),
     install_requires=[
         'terminaltables',
@@ -20,8 +20,8 @@ setup(
     license='GNU General Public License v3 (GPLv3) (GPL)',
     zip_safe=True,
     entry_points={
-        'console_scripts':[
-            'drugwars = drugwars.__main__:main',
+        'console_scripts': [
+            'agentics = agentics.__main__:main',
             'stockwolf = stockwolf.main:main'
         ],
     },
